functions/mention_tracker/requirements.txt
functions-framework==3.5.0
google-cloud-bigquery==3.14.1
psycopg2-binary==2.9.9
